refactor(#93189): Extracting "XsdDateTimeParser" and it's environments to separate files.
