+++ src/libraries/System.Private.Xml/src/System.Private.Xml.csproj
@@ -42,6 +42,9 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>TextUtf8RawTextWriter.tt</DependentUpon>
     </Compile>
+    <Compile Include="System\Xml\Schema\DateAndTime\Specifications\DateTimeTypeCode.cs" />
+    <Compile Include="System\Xml\Schema\DateAndTime\Parsers\XsdDateTimeParser.cs" />
+    <Compile Include="System\Xml\Schema\DateAndTime\Specifications\XsdDateTimeKind.cs" />
     <None Include="System\Xml\Core\HtmlEncodedRawTextWriter.tt">
       <Generator>TextTemplatingFileGenerator</Generator>
       <LastGenOutput>HtmlEncodedRawTextWriter.cs</LastGenOutput>
