+++ src/libraries/System.Private.Xml/src/System/Xml/Schema/DateAndTime/Specifications/DateTimeTypeCode.cs
@@ -1,0 +1,21 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace System.Xml.Schema.DateAndTime.Specifications
+{
+    /// <summary>
+    /// Subset of represented XML Schema types.
+    /// </summary>
+    internal enum DateTimeTypeCode
+    {
+        DateTime,
+        Time,
+        Date,
+        GYearMonth,
+        GYear,
+        GMonthDay,
+        GDay,
+        GMonth,
+        XdrDateTime,
+    }
+}
